refactor: extract Excalidraw extension renderer
